requirements.txt
requirements_txt = """
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
sqlite3
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.0

# AI and ML
google-generativeai==0.3.2
langchain==0.0.350
langchain-community==0.0.3
sentence-transformers==2.2.2
chromadb==0.4.18

# Document processing
PyMuPDF==1.23.8
pytesseract==0.3.10
Pillow==10.1.0
python-docx==1.1.0

# HTTP client
httpx==0.25.2
aiohttp==3.9.1

# Logging
structlog==23.2.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
"""